Marked some tests from python regression test suite as 'skipped'.
Some tests use features not supported under ZeroVM, such as pipes,
subprocessing. These tests are not supposed to pass, so mark them to be
skipped.

We have platform define as "linux2-zvm" and using this condition to check
if tests should be skipped.

This will make tests pass. Next step is to rewrite some of them. This would 
be done outside python regression test suite.